Fix MOBI-2140
Exchange and getaways - search results - top scrolling calendar is working incorrectly
This is not happening on Andoroid. 
Steps:
login (Geekosis)
select vacation search 
scottsdale for August 21 
the scrolling caledar is not working correctly.
It is showing you a combination of no availability screen and availability screen. 
It is taking you to the front or end of the calendar when selecting a month bucket
it showing you no availability but does not have the date/month grouping in view
Attached are two video examples.